requirements.txt
discord.py>=1.6<2.0.0
python-dotenv
fuzzywuzzy[speedup]